Add support of dimming of backlight.
Feature gets value of acpi backlight value and sets user defined value.
By default, I decided to leave configurations values commented as in
testing.  I've known a good number of people to use this with success
but am unwilling to attest to just yet.  Have test setting and the
screen does dim on battery and return to set value on ac.  From the
ThinkWiki I've read that this can save a decent amount of power.