Issues #147, #149: fix udev event processing for AC/BAT switching
Background: systemd-udevd version 221 now kills forked processes
when the main process (called by RUN=) terminates and stdout/
stderr have been closed. First occurred on Arch Linux.

Consequences:
TLP's technique to fork udev event processing in a subshell fails
and the code applying the settings is aborted prematurely.

In case of tlp-usb-udev removing the subshell is not an option,
because it would remove the ability to concurrently populate
the subdevices tree and thus incapacitate the driver blacklisting
feature.

Solution:
- Remove subshell forking for AC/BAT switching in tlp
- Do not close stdout/stdin in tlp-usb-udev subshell to prevent
  being killed by udev post-processing
- Reduce wait for subdevices populating in tlp-usb-udev