tlp.service: Type=oneshot
As is known there is no TLP daemon but only a start/stop script, so 'oneshot' is
better suited than 'simple' which is intended for true daemons.

Arch wiki [1] suggests "Type=oneshot: this is useful for scripts that do a single
job and then exit."

systemd.service manpage states "Behavior of oneshot is similar to simple; however,
it is expected that the process has to exit before systemd starts follow-up units.
RemainAfterExit= is particularly useful for this type of service."

[1] https://wiki.archlinux.org/index.php/Systemd#Service_types