Fix rdw_nm script to wait in case of timed lock
If NetworkManager run the dispatcher scripts too fast, the rdw_nm won't
be able to turn off the required devices, e.g., NetworkManager turns on
the WiFi fist and then, the Ethernet link. Normally the user will configure TLP
so it turns off the Wifi, if the Ethernet link is on. The first execution
of rdw_nm won't do anything but sets up a timed lock. The second
execution should turn off the WiFi, but since there's a valid timed lock, it
exits immediately leaving the network devices in the wrong state.

With this patch, rdw_nm will wait at least three times in case there's a
valid timed lock.