Makefile: sed INFILES in parallel
Eliminate the shell FOR-LOOP and use an equivalent
Makefile construct that plays nicely with "make -j".

This also allows to sed single files,
e.g. "make tlp-functions".

Contrary to the shell FOR-LOOP, this variant does not
recreate existing files unless their ".in" file changed,
i.e. changes to TLP_SBIN, ... are not considered
when make decides which targets need to be rebuild.
With this change, the install targets do no longer
trigger a rebuild,
which previously resulted in 2 builds when
running "make install-rdw" after "make install".

Signed-off-by: André Erdmann <dywi@mailerd.de>
Signed-off-by: Thomas Koch <linrunner@gmx.net>