Add test for presence of /var/run
If /var is mounted from a different partition, the /var/run directory may not be present immediately. In that case, /var/run/tlp should not be created, but rather we should wait until it is present.

Otherwise, the /var mountpoint in the / filesystem will not be empty (which is wrong) and anything below that won't be reachable anyway once /var (and /var/run) is mounted. Adding a test for the presence of /var/run will prevent this from happening.