read_config(): get rid of tlp subdir in non-root mode
Using a subdirectory of tmpdir for storing
the temporary configuration file conflicts with multi-user systems:

Regular user A runs a TLP command
that creates a temporary configuration file, such as "tlp-stat --config".
Assuming the default tmpdir /tmp/tlp and a standard umask of 0022,
"/tmp/tlp" will belong to that user with mode 0755.
Any other non-root user will not be able to write their
temporary configuration file to that directory since mktemp will fail.

So, do not use a subdir and write to $TMPDIR instead,
assuming that it is world-writable (e.g. mode=1777).
To keep the config file distinguishable, prepend "tlp-" to the filename.

Signed-off-by: Thomas Koch <linrunner@gmx.net>