packaging: make TLPVER a build-time variable
Gentoo packages make use of TLPVER to indicate
(a) live builds (=git master as of time of package build)
    by appending the date / git sha ref to TLPVER
(b) unsupported builds, e.g. when unbundling tpacpi-bat

Currently, this is handled by the tlp-gentoo-additions repo,
which uses some sed to edit the tlp functions file in-place.

To make this more convenient, this commit suggests storing the
version string in the "VERSION" file where the Makefile
can catch it up by reading the first word of that file.

In future, this could also be used to generate
the version string in the man page files.

Signed-off-by: Thomas Koch <linrunner@gmx.net>