packaging: make TPACPIBAT prog path a build-time variable
In an effort of unbundling dependencies, the Gentoo package allows users
to use the upstream tpacpi-bat (app-laptop/tpacpi-bat),
which gets installed in a systemd-wide "/bin" directory (usually /usr/bin/).

This is highly unsupported and has lead to issues in the past (linrunner/TLP#44),
but is an option for those who follow tpacpi-bat and TLP development closely
and/or do not want to have that extra copy of tpacpi-bat for whatever reason.

As a side effect, this commit also allows installing the bundled tpacpi-bat
to a different location, e.g. "make TPACPIBAT=/usr/bin/tpacpi-bat".

Signed-off-by: Thomas Koch <linrunner@gmx.net>