tlp-sleep.service: reintroduce
This patch reintroduces tlp-sleep as a systemd unit as opposed to a
systemd system-sleep hook script. This is due to the recommendation by
systemd itself to not use the hook scripts. As per the manual:

> Note that scripts or binaries dropped in /usr/lib/systemd/system-sleep/
> are intended for local use only and should be considered hacks. If
> applications want to react to system suspend/hibernation and resume,
> they should rather use the Inhibitor interface[1].

This also fixes an issue with the previous unit where if `/var/lib/tlp`
did not exist it would fail. By addind a `StateDirectory` clause we now
let systemd handle the creation of the dir.

c.f. `man systemd-sleep`