tlp-stat -b: map battery state 'Unknown' to 'Idle' for clarity
Rationale: status 'Unknown' is puzzling for many users.
In case of ThinkPads with charge thresholds it just means 'Idle
because of a threshold'.

Other laptop brands with natacpi/thresholds may run into the same
problem because kernel devs refuse to return 'Idle' or 'Not charging'
state as long as upowerd chokes on it.

Solution:

* 'Unknown' is now 'Idle' - on AC only
* Drop the cryptic 'threshold may prevent charging' hint,
  it may be inaccurate e.g. in case of a dual battery ThinkPad
* Fix broken detection of power source too