bat.d (asus): apply stop threshold reliably on boot and after hibernate
Rationale: asus_wmi.c states that the actual currently used value for the
threshold cannot be read from the firmware. Thus, a cached valued is used
to read the value back, which does not coincide with the actual value
in the firmware after a resume from hibernate.

Solution: remove the read-and-compare code, always write the thresholds.
As a precaution, this is implemented for all threshold write operations,
i.e. boot, resume from suspend/hibernate and command line.

References:
* https://github.com/linrunner/TLP/issues/589
* https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/drivers/platform/x86/asus-wmi.c#n640